Split Protocol into its own source file
