+++ lib/go-tc/enum/enum.go
@@ -31,7 +31,6 @@ package enum
 import (
 	"encoding/json"
 	"errors"
-	"fmt"
 	"strconv"
 	"strings"
 )
@@ -126,80 +125,6 @@ func DSTypeCategoryFromString(s string) DSTypeCategory {
 
 const SigningAlgorithmURLSig = "url_sig"
 const SigningAlgorithmURISigning = "uri_signing"
-
-const DSProtocolHTTP = 0
-const DSProtocolHTTPS = 1
-const DSProtocolHTTPAndHTTPS = 2
-const DSProtocolHTTPToHTTPS = 3
-
-// Protocol represents an ATC-supported content delivery protocol.
-type Protocol string
-
-const (
-	// ProtocolHTTP represents the HTTP/1.1 protocol as specified in RFC2616.
-	ProtocolHTTP = Protocol("http")
-	// ProtocolHTTPS represents the HTTP/1.1 protocol over a TCP connection secured by TLS
-	ProtocolHTTPS = Protocol("https")
-	// ProtocolHTTPtoHTTPS represents a redirection of unsecured HTTP requests to HTTPS
-	ProtocolHTTPtoHTTPS = Protocol("http to https")
-	// ProtocolHTTPandHTTPS represents the use of both HTTP and HTTPS
-	ProtocolHTTPandHTTPS = Protocol("http and https")
-	// ProtocolInvalid represents an invalid Protocol
-	ProtocolInvalid = Protocol("")
-)
-
-// String implements the "Stringer" interface.
-func (p Protocol) String() string {
-	switch p {
-	case ProtocolHTTP:
-		fallthrough
-	case ProtocolHTTPS:
-		fallthrough
-	case ProtocolHTTPtoHTTPS:
-		fallthrough
-	case ProtocolHTTPandHTTPS:
-		return string(p)
-	default:
-		return "INVALIDPROTOCOL"
-	}
-}
-
-// ProtocolFromString parses a string and returns the corresponding Protocol.
-func ProtocolFromString(s string) Protocol {
-	switch strings.Replace(strings.ToLower(s), "_", " ", -1) {
-	case "http":
-		return ProtocolHTTP
-	case "https":
-		return ProtocolHTTPS
-	case "http to https":
-		return ProtocolHTTPtoHTTPS
-	case "http and https":
-		return ProtocolHTTPandHTTPS
-	default:
-		return ProtocolInvalid
-	}
-}
-
-// UnmarshalJSON implements the json.Unmarshaler interface.
-func (p *Protocol) UnmarshalJSON(data []byte) error {
-	if string(data) == "null" {
-		return errors.New("Protocol cannot be null")
-	}
-	s, err := strconv.Unquote(string(data))
-	if err != nil {
-		return fmt.Errorf("JSON %s not quoted: %v", data, err)
-	}
-	*p = ProtocolFromString(s)
-	if *p == ProtocolInvalid {
-		return fmt.Errorf("%s is not a (supported) Protocol", s)
-	}
-	return nil
-}
-
-// MarshalJSON implements the json.Marshaler interface.
-func (p Protocol) MarshalJSON() ([]byte, error) {
-	return json.Marshal(p.String())
-}
 
 // DeepCachingType represents a Delivery Service's deep caching type. The string values of this type should match the Traffic Ops values.
 type DeepCachingType string
